Refactor and code styling
